Struct la::CholeskyDecomposition

pub struct CholeskyDecomposition<T> {
    // some fields omitted
}

Cholesky Decomposition (for a real valued matrix).

Initial implementation based on JAMA.

For a symmetric, positive definite matrix A, the Cholesky decomposition is an lower triangular matrix L so that A = L*L'.

If the matrix is not symmetric or positive definite, None is returned.

Methods

impl<T: Float + ApproxEq<T>> CholeskyDecomposition<T>

fn new(m: &Matrix<T>) -> Option<CholeskyDecomposition<T>>

fn get_l<'lt>(&'lt self) -> &'lt Matrix<T>

fn solve(&self, b: &Matrix<T>) -> Matrix<T>