Struct la::EigenDecomposition

pub struct EigenDecomposition<T> {
    // some fields omitted
}

Eigenvalues and eigenvectors of a real matrix.

Ported from JAMA.

If A is symmetric, then A = V*D*V' where the eigenvalue matrix D is diagonal and the eigenvector matrix V is orthogonal. I.e. A = V * D * V' and V * V' = I.

If A is not symmetric, then the eigenvalue matrix D is block diagonal with the real eigenvalues in 1-by-1 blocks and any complex eigenvalues, lambda + i*mu, in 2-by-2 blocks, [lambda, mu; -mu, lambda]. The columns of V represent the eigenvectors in the sense that A*V = V*D, The matrix V may be badly conditioned, or even singular, so the validity of the equation A = V * D * V-1 depends upon V.cond().

Methods

impl<T: Float + ApproxEq<T> + Signed> EigenDecomposition<T>

fn new(a: &Matrix<T>) -> EigenDecomposition<T>

fn get_v<'lt>(&'lt self) -> &'lt Matrix<T>

fn get_real_eigenvalues<'lt>(&'lt self) -> &'lt Vec<T>

fn get_imag_eigenvalues<'lt>(&'lt self) -> &'lt Vec<T>

fn get_d(&self) -> Matrix<T>