Struct la::QRDecomposition

pub struct QRDecomposition<T> {
    // some fields omitted
}

QR Decomposition.

Based on Apache Commons Math and JAMA.

For an m-by-n matrix A, the QR decomposition is an m-by-m orthogonal matrix Q and an m-by-n upper triangular (or trapezoidal) matrix R, so that A = Q*R.

The QR decompostion always exists, even if the matrix does not have full rank. The primary use of the QR decomposition is in the least squares solution of nonsquare systems of simultaneous linear equations. This will fail if is_full_rank() returns false.

Methods

impl<T: Float + ApproxEq<T>> QRDecomposition<T>

fn new(m: &Matrix<T>) -> QRDecomposition<T>

fn is_full_rank(&self) -> bool

fn get_h(&self) -> Matrix<T>

fn get_r(&self) -> Matrix<T>

fn get_q(&self) -> Matrix<T>

fn solve(&self, b: &Matrix<T>) -> Option<Matrix<T>>