Struct la::QRDecomposition

pub struct QRDecomposition<T> { /* fields omitted */ }

QR Decomposition.

Based on Apache Commons Math and JAMA.

For an m-by-n matrix A, the QR decomposition is an m-by-m orthogonal matrix Q and an m-by-n upper triangular (or trapezoidal) matrix R, so that A = Q*R.

The QR decompostion always exists, even if the matrix does not have full rank. The primary use of the QR decomposition is in the least squares solution of nonsquare systems of simultaneous linear equations. This will fail if is_full_rank() returns false.

Methods

impl<T: Float + ApproxEq<T>> QRDecomposition<T>

pub fn new(m: &Matrix<T>) -> QRDecomposition<T>

pub fn is_full_rank(&self) -> bool

pub fn get_h(&self) -> Matrix<T>

pub fn get_r(&self) -> Matrix<T>

pub fn get_q(&self) -> Matrix<T>

pub fn solve(&self, b: &Matrix<T>) -> Option<Matrix<T>>